+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 ### 5.1.21
+
++ [x] Fixed method `getTabByTitle()` in `page-control`, added prop `hasMenu: $el.attr("data-menu") !== "false"` to method `createStruct()` (issues 2007, 2008)